feat: Subscribe to multiple topics (API only) (#119)
